requirements.txt
djangorestframework==3.13.1
djangorestframework-camel-case==1.1.2
entrypoints==0.3
requests==2.27
gunicorn==20.1.0
django-cors-headers==3.11.0
pymemcache==3.5.0
django_redis==5.2.0
rq==1.10.1
celery==5.2.6